requirements.txt
requests>=2.31.0
yfinance>=0.2.36
finnhub-python>=2.4.18
google-generativeai>=0.4.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
